Moved the core algorithm to a different file.
